update xml/xsd for unsigned updates. updating tests to reflect this...

+++ jobs/src/main/java/org/openstack/atlas/atom/mapper/LbaasUsageDataMapper.java
@@ -24,6 +24,7 @@ import org.w3._2005.atom.UsageContent;
 
 import javax.ws.rs.core.MediaType;
 import javax.xml.datatype.DatatypeConfigurationException;
+import java.math.BigInteger;
 import java.security.NoSuchAlgorithmException;
 import java.util.UUID;
 
@@ -105,10 +106,10 @@ public class LbaasUsageDataMapper {
         LBaaSUsagePojo lu = new LBaaSUsagePojo();
         lu.setAvgConcurrentConnections(usageRecord.getAverageConcurrentConnections());
         lu.setAvgConcurrentConnectionsSsl(usageRecord.getAverageConcurrentConnectionsSsl());
-        lu.setBandWidthOutSsl(usageRecord.getOutgoingTransferSsl());
-        lu.setBandWidthInSsl(usageRecord.getIncomingTransferSsl());
-        lu.setBandWidthOut(usageRecord.getOutgoingTransfer());
-        lu.setBandWidthIn(usageRecord.getIncomingTransfer());
+        lu.setBandWidthOutSsl(BigInteger.valueOf(usageRecord.getOutgoingTransferSsl()));
+        lu.setBandWidthInSsl(BigInteger.valueOf(usageRecord.getIncomingTransferSsl()));
+        lu.setBandWidthOut(BigInteger.valueOf(usageRecord.getOutgoingTransfer()));
+        lu.setBandWidthIn(BigInteger.valueOf(usageRecord.getIncomingTransfer()));
         lu.setResourceType(ResourceTypes.LOADBALANCER);
         lu.setNumPolls(usageRecord.getNumberOfPolls());
         lu.setNumVips(usageRecord.getNumVips());
